ontology_release/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aimworks-ontology-release"
version = "2026.3.0"
description = "Release preparation pipeline for the H2KG PEMFC Catalyst Layer Application Ontology"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AIMWORKS Ontology Engineering Team" }]
dependencies = [
  "rdflib>=7.0.0",
  "rdflib-jsonld>=0.6.2",
  "PyYAML>=6.0.1",
  "requests>=2.32.0",
  "pandas>=2.2.0",
  "rapidfuzz>=3.9.0",
  "pyshacl>=0.26.0",
  "jinja2>=3.1.4",
  "typer>=0.12.3",
  "pytest>=8.2.0"
]

[project.optional-dependencies]
emmo = ["EMMOntoPy>=0.6.0", "owlready2>=0.46"]

[project.scripts]
aimworks-ontology-release = "aimworks_ontology_release.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q --basetemp=.pytest_tmp -p no:cacheprovider"
